requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dateutil
fastmcp